% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocs4R.R
\docType{package}
\name{ocs4R}
\alias{ocs4R}
\alias{ocs4R-package}
\title{Interface to 'OCS' REST API}
\description{
Provides an Interface to 'OCS' (<https://www.open-collaboration-services.org/>) REST API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/eblondel/ocs4R}
  \item Report bugs at \url{https://github.com/eblondel/ocs4R/issues}
}

}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
