% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocs4R.R
\docType{package}
\name{ocs4R}
\alias{ocs4R}
\alias{ocs4R-package}
\title{Interface to 'OCS' REST API}
\description{
Provides an Interface to 'OCS' (<http://www.open-collaboration-services.org/>) REST API.
}
\details{
\tabular{ll}{
   Package: \tab ocs4R\cr
   Type: \tab Package\cr
   Version: \tab 0.1-1\cr
   Date: \tab 2020-05-28\cr
   License: \tab MIT\cr
   LazyLoad: \tab yes\cr
 }
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
