% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocsApiUserProvisioningManager.R
\docType{class}
\name{ocsApiUserProvisioningManager}
\alias{ocsApiUserProvisioningManager}
\title{ocsApiUserProvisioningManager}
\format{\code{\link{R6Class}} object.}
\usage{
ocsApiUserProvisioningManager
}
\value{
Object of \code{\link{R6Class}} for modelling an ocsManager for Webdav API
}
\description{
ocsApiUserProvisioningManager
}
\section{General Methods (inherited from 'ocsManager')}{

\describe{
 \item{\code{new(url, user, pwd, logger)}}{
   This method is used to instantiate an ocsApiUserProvisioningManager. The user/pwd are
   mandatory in order to connect to 'ocs'. The logger can be either
   NULL, "INFO" (with minimum logs), or "DEBUG" (for complete curl 
   http calls logs)
 }
 \item{\code{connect()}}{
   A method to connect to 'ocs' and set version/capabilities
 }
 \item{\code{getVersion()}}{
   Get the 'ocs' server version
 }
 \item{\code{getCapabilities()}}{
   Get the 'ocs' server capabilities
 }
}
}

\section{User Provisioning API methods}{

\describe{
 \item{\code{addUser(userid, email, password, groups)}}{
   Adds a user given a \code{userid} (required). All other fields (email, password, groups) are
   optional for the user creation. Returns \code{TRUE} if the user is added, \code{FALSE} otherwise.
 }
 \item{\code{getUsers()}}{
   Get the list of users. This method returns a vector of class 'character' giving
   the user IDs available in the OCS cloud plateform.
 }
 \item{\code{getUser(userid, pretty)}}{
   Get the user details from its \code{userid}. If the argument \code{pretty} is set to TRUE,
   this will return an object of class \code{data.frame}, otherwise (by default) it returns 
   an object of class \code{list}.
 }
 \item{\code{editUser(userid, key, value)}}{
   Edits a user, identifier by a userid. The user property to be edited should be set using its
   key (eg displayname) and the value to be modified for this key. Returns \code{TRUE} if the user 
   is edited, \code{FALSE} otherwise.
 }
 \item{\code{enableUser(userid)}}{
   Enables a user. Returns \code{TRUE} if enabled.
 }
 \item{\code{disableUser(userid)}}{
   Disables a user. Returns \code{TRUE} if disabled.
 }
 \item{\code{deleteUser(userid)}}{
   Deletes auser. Returns \code{TRUE} if delete.
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{api}
\keyword{manager}
\keyword{ocs}
\keyword{userprovisioning}
