% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocsRequest.R
\docType{class}
\name{ocsRequest}
\alias{ocsRequest}
\title{ocsRequest}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a generic 'ocs' web-service request
}
\description{
ocsRequest
}
\note{
Abstract class used internally by \pkg{ocs4R}
}
\section{Methods}{

\describe{
 \item{\code{new(type, url, request, user, pwd, logger)}}{
   This method is used to instantiate a object for doing an 'ocs' web-service request
 }
 \item{\code{getRequest()}}{
   Get the request payload
 }
 \item{\code{getRequestHeaders()}}{
   Get the request headers
 }
 \item{\code{getStatus()}}{
   Get the request status code
 }
 \item{\code{getResponse()}}{
   Get the request response
 }
 \item{\code{getException()}}{
   Get the exception (in case of request failure)
 }
 \item{\code{getResult()}}{
   Get the result \code{TRUE} if the request is successful, \code{FALSE} otherwise
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ocs}
\keyword{request}
